Refactor MCP DTO normalization helpers
